emacs: switch back to org-mode
After using a plain init.el for quite a while, I've found that one of
the pain points relates around organizing language settings. I was
making blocks for python-mode (as an example) which contained settings
for a number of python related packages, even though there were no other
settings in the python-mode block. This worked fine, but it seems like
org-mode might be a smarter option for this, as it has headings and
stuff like that built in.