nix: reorganize files, drop some mkHome magic
The simple explanation is that it makes more sense to me to have all
related files in subfolders. Previously, things were organized into a
sort of "separation of concerns-ish" way, but that meant that
home-manager modules were next to nixos modules, and it just got
confusing to manage.

This change creates top-level folders for each type of deployment
(NixOS, nix-darwin, home-manager). This allows all the logic for that
type of target to be in a single top-level folder - we're not looking at
`modules/nixos/default.nix` and `hosts/nixos/zagreus`, we're looking at
`nixos/modules/default.nix` and `nixos/hosts/zagreus`.

This may seem like a minor change, but with how often I had to second
guess where a file was, it just made sense.